refactor environment transformation
