+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/environment/EnvironmentTransformationPipeline.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.environment
 
 import com.spotify.scio.ScioResult
 import org.broadinstitute.monster.common.{PipelineBuilder, ScioApp}
