+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/environment/EnvironmentTransformations.scala
@@ -1,6 +1,6 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.environment
 
-import org.broadinstitute.monster.dap.environment._
+import org.broadinstitute.monster.dap.InvalidArmMonthError
 import org.broadinstitute.monster.dap.HLESurveyTransformationPipelineBuilder.logger
 import org.broadinstitute.monster.dap.common.RawRecord
 import org.broadinstitute.monster.dogaging.jadeschema.table.Environment
